backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.10.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.7.1
resend==2.5.0
python-dotenv==1.0.1
python-multipart==0.0.20
groq==0.13.0
